backend/requirements.txt
alembic==1.18.4
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
PyJWT==2.11.0
python-dotenv==1.2.2
requests==2.32.5
SQLAlchemy==2.0.48
tomli==2.4.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
